Added DropZone.js for file uploads + removed showFile button on feedback

+++ src/app/app/routes/feedback.js
@@ -9,7 +9,17 @@ var fm = require('../fileMarker.js');
 
 /* GET feedback page. */
 router.get('/', function(req, res, next) {
-  res.redirect('/');
+    
+    //If template variables exist, DropZone is finished so render page
+    if (req.app.locals.status) {
+        res.render('feedback', { title: 'Upload Feedback', file: req.app.locals.fileName, status: req.app.locals.status, passRate:req.app.locals.passRate, contents: req.app.locals.fileContents});
+    }
+
+    //If no variables exist, user is making plain GET request to feedback, therefore redirect to root
+    else {
+        res.redirect('/');
+    }
+  
 
 });
 
@@ -33,12 +43,13 @@ router.post('/', function(req, res, next) {
     var email = req.app.locals.emailAddr
     var user = email.substring(0, email.lastIndexOf('@')); 
     var names = user.split('.');
+    var userName = names[0] + names[1];
 
     //Save to uploads directory
     form.on('fileBegin', function (name, file){
         if (file.name != ""){
 
-            tmpDir = path.join(__dirname, '../uploads/' + names[0] + names[1] + '/');
+            tmpDir = path.join(__dirname, '../uploads/' + userName + '/');
             if(!fs.existsSync(tmpDir)){
                 fs.mkdirSync(tmpDir);
             }
@@ -65,13 +76,16 @@ router.post('/', function(req, res, next) {
 
                 //Create directory in student data 
                 var studentDir = path.join(__dirname, '../studentdata/'); 
-                var userSaveDir = studentDir + names[0] + names[1];
+                var userSaveDir = studentDir + userName + '/';
 
                 //If file is uploaded, render feedback page and clear timeout
                 if (fs.existsSync(fileLocation)) {
 
+                    //Make file executable
+                    fs.chmodSync(fileLocation, 777);
+
                     //Call file marker code and find out whether correct/incorrect
-                    fm.checkForMarker(file, function(status, passRate){
+                    fm.checkForMarker(file, userName, function(status, passRate){
 
                         //If file is valid
                         if (status != "invalid"){
@@ -84,7 +98,7 @@ router.post('/', function(req, res, next) {
                             });
 
                             //Path to where filed is saved
-                            userSavePath = userSaveDir + '/' + fileName;
+                            userSavePath = userSaveDir + fileName;
 
                             //Check if user directory exists, create if not
                             if (!fs.existsSync(userSaveDir)){  
@@ -98,24 +112,31 @@ router.post('/', function(req, res, next) {
                                 }
                             });
 
-                            fs.unlinkSync(fileLocation);
-                        }
-                        
-                        if (status == 'correct') {
-                            res.render('feedback', { title: 'Upload Feedback', file: fileName, status: 'Correct', passRate:passRate, contents: fileContents});
-                        }
+                            var fileCopiedRetry = setInterval(fileCopied, 1000);
 
-                        else if (status == 'incorrect') {
-                            res.render('feedback', { title: 'Upload Feedback', file: fileName, status: 'Incorrect', passRate:passRate, contents: fileContents});
-                        }
+                            //Wait until file has been copied
+                            function fileCopied(){
 
-                        else {
-                            res.render('feedback', { title: 'Upload Feedback', file: fileName, status: 'Invalid upload file name'});
+                                //When file has been copied, delete from uploads
+                                if (fs.existsSync(userSavePath)){
+                                
+                                    fs.chmodSync(userSavePath, 777);
+                                    fs.unlinkSync(fileLocation);
+
+                                    //Set variables for template render
+                                    req.app.locals.fileName = fileName;
+                                    req.app.locals.status = status;
+                                    req.app.locals.passRate = passRate;
+                                    req.app.locals.fileContents = fileContents;
+                                    
+                                    clearInterval(fileCopiedRetry);
+                                    return res.status(200).send('');
+                                };
+                            }
                         }
+                    });
 
                     clearInterval(uploadTimeout);
-
-                    });
                 }
             }      
         }
